fix warning message for __get_libvec_base

+++ src/zos.cc
@@ -2568,7 +2568,7 @@ int *__get_stack_start() {
 }
 
 unsigned long __get_libvec_base() {
-  int psalaa = ((int* __ptr32)(1208))[0];
+  unsigned long psalaa = ((int* __ptr32)(1208))[0];
   unsigned long lca64 = *(unsigned long*)(psalaa + 88);
   unsigned long caa64 = *(unsigned long*)(lca64 + 8);
   return *(unsigned long*)(caa64 + 1016);
